[#32] optimize deployNode.sh and deployNode.py files
