+++ deployNode/deployNode.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+nohup python3 ./deployNode.py 2>&1 >/dev/null &
